fix: improve handling of invitations
ensure that space & device invitations are always handled (particularly
in dev mode/when auth is skipped) even when there is no existing
identity
